requirements.txt
python-multipart>=0.0.6
fastapi==0.95.0
uvicorn==0.21.1
streamlit==1.22.0
numpy==1.24.3
clip-anytorch==2.5.0
opencv-python-headless==4.7.0.72
requests==2.28.2
firebase-admin==6.1.0
python-dotenv==1.0.0
plotly==5.15.0
scikit-learn==1.2.2
gunicorn==20.1.0
google-cloud-run==0.7.1
torch